Fix theme issue with smartbar picker dialog
This clears up a white on white issue where the 'fab' behind the actions
in the smartbar picker dialog adapts to the system accent, in this case
being white and the action items are also white. Since white is the trouble
maker of all the accents, we're going to just change the color of the 'fab'
when is enabled. Thus leaving the look that was picked by design in place
for the other accents. Win win.

Change-Id: I3606a8776a446a4a6a099a54e7546ab3ba250c5a
Signed-off-by: spezi77 <spezi77@gmx.com>